pyproject.toml
[project]
name = "toxiscore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
